delete

+++ src/main/java/com/ll/auth/domain/post/post/controller/ApiV1PostController.java
@@ -41,8 +41,19 @@ public class ApiV1PostController {
     }
 
     @DeleteMapping("/{id}")
-    public RsData<Void> deleteItem(@PathVariable long id) {
+    public RsData<Void> deleteItem(@PathVariable long id, long authorId, String password) {
+        Member actor = memberService.findById(authorId).get();
+
+        if (!actor.getPassword().equals(password)) {
+            throw new ServiceException("401-1", "비밀번호가 일치하지 않습니다.");
+        }
+
         Post post = postService.findById(id).get();
+
+        if (!post.getAuthor().equals(actor)) {
+            throw new ServiceException("403-1", "작성자만 글을 삭제할 권한이 있습니다.");
+        }
+
         postService.delete(post);
 
         return new RsData<>(
